feat: singleton toast
